% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Align templates and images}
\usage{
align(
  stimlist,
  pt1 = 0,
  pt2 = 1,
  x1 = NULL,
  y1 = NULL,
  x2 = NULL,
  y2 = NULL,
  width = NULL,
  height = NULL,
  ref_img = 1,
  fill = "white",
  patch = FALSE,
  squash = FALSE,
  procrustes = FALSE
)
}
\arguments{
\item{stimlist}{list of class webmorph_list}

\item{pt1}{The first point to align (defaults to 0)}

\item{pt2}{The second point to align (defaults to 1)}

\item{x1}{The x-coordinate to align the first point to}

\item{y1}{The y-coordinate to align the first point to}

\item{x2}{The x-coordinate to align the second point to}

\item{y2}{The y-coordinate to align the second point to}

\item{width}{The width of the aligned image}

\item{height}{The height of the aligned image}

\item{ref_img}{The reference image (by index or name) to get coordinates and dimensions from if NULL (defaults to first image)}

\item{fill}{background color if cropping goes outside the original image}

\item{patch}{whether to use the patch function to set the background color}

\item{squash}{whether to move template points outside the image boundaries inside the image}

\item{procrustes}{Whether to do a procrustes alignment}
}
\value{
webmorph_list with aligned tems and/or images
}
\description{
Align templates and images
}
\examples{
faces("test") \%>\%
  align(x1 = 100, y1 = 100, x2 = 200, y2 = 200,
        width = 300, height = 400) \%>\%
  plot()

 faces("lisa")[1:2] \%>\%
   align(procrustes = TRUE) \%>\%
   plot()
}
