% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_webmorph.R
\name{assert_webmorph}
\alias{assert_webmorph}
\title{Check for webmorph_list}
\usage{
assert_webmorph(x)
}
\arguments{
\item{x}{The object}

\item{arg}{The name of the argument}
}
\value{
A webmorph_list
}
\description{
Check if an object is a list of class webmorph_list. If class webmorph_stim, wrap in webmorph_list. If a properly structured list without the right class, add the right class.
}
