% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{batchTrans}
\alias{batchTrans}
\title{Batch Transform}
\usage{
batchTrans(
  batch,
  project_id,
  norm = c("none", "twopoint", "rigid"),
  normpoint = 0:1,
  sample_contours = TRUE,
  warp = c("multiscale", "linear", "multiscalerb"),
  format = c("jpg", "png", "gif")
)
}
\arguments{
\item{batch}{data frame containing batch info}

\item{project_id}{the project ID}

\item{norm}{how to normalise}

\item{normpoint}{points for twopoint normalisation}

\item{sample_contours}{whether to sample contours}

\item{warp}{warping algorithm to use}

\item{format}{image format}
}
\value{
webmorph_list
}
\description{
Batch Transform
}
