% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bounds}
\alias{bounds}
\title{Get template bounds}
\usage{
bounds(stimlist, each = FALSE)
}
\arguments{
\item{stimlist}{A webmorph_list}

\item{each}{Whether to calculate max and min for the full set (default) or each image separately}
}
\value{
A list of min and max x and y values
}
\description{
Get template bounds
}
\examples{
faces("london") \%>\% bounds()

faces("test") \%>\% bounds(each = TRUE)
}
