% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c.webmorph_list}
\alias{c.webmorph_list}
\title{Combine webmorph_lists}
\usage{
\method{c}{webmorph_list}(...)
}
\arguments{
\item{...}{webmorph_lists to be concatenated}
}
\value{
webmorph_list
}
\description{
Combine webmorph_lists
}
