% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c.webmorph_stim}
\alias{c.webmorph_stim}
\title{Combine webmorph_stim}
\usage{
\method{c}{webmorph_stim}(...)
}
\arguments{
\item{...}{webmorph_stim to be concatenated}
}
\value{
webmorph_list
}
\description{
Combine webmorph_stim
}
