% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c.webmorph_tem}
\alias{c.webmorph_tem}
\title{Combine templates}
\usage{
\method{c}{webmorph_tem}(...)
}
\arguments{
\item{...}{webmorph templates to be concatenated}
}
\value{
webmorph template list
}
\description{
Combine templates
}
