% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c.webmorph_temlist}
\alias{c.webmorph_temlist}
\title{Combine template lists}
\usage{
\method{c}{webmorph_temlist}(...)
}
\arguments{
\item{...}{webmorph template lists to be concatenated}
}
\value{
webmoprh template list
}
\description{
Combine template lists
}
