% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_when.R
\name{case_when_jit}
\alias{case_when_jit}
\title{case_when with JinT Evaluation}
\usage{
case_when_jit(...)
}
\arguments{
\item{...}{formulae to evaluate}
}
\value{
first true value or NULL
}
\description{
Works like dplyr::case_when, but RHS doesn't all need to be the same data type, RHS isn't evaluated in LHS is FALSE, and LHE isn't evaluated unless all previous LHS are FALSE
}
\examples{
case_when_jit(F ~ "A", T ~ 1.1, bad_val ~ "B")
}
