% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_temlist.R
\name{check_temlist}
\alias{check_temlist}
\title{Check for temlist}
\usage{
check_temlist(x, arg = "temlist")
}
\arguments{
\item{x}{The object}

\item{arg}{The name of the argument}
}
\value{
A temlist
}
\description{
Check if an object is a temlist. If a single tem, convert to a temlist.
}
