% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{dirDelete}
\alias{dirDelete}
\title{Delete directories on webmorph.org}
\usage{
dirDelete(dir)
}
\arguments{
\item{dir}{directory to delete}
}
\value{
logical, if directory was deleted
}
\description{
Delete directories on webmorph.org
}
