% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{dirLoad}
\alias{dirLoad}
\title{Get directory contents}
\usage{
dirLoad(dir = "", project_id = Sys.getenv("webmorph_project_id"))
}
\arguments{
\item{dir}{the directory to look in}

\item{project_id}{the project ID}
}
\value{
nested list of the directory
}
\description{
Get directory contents
}
