% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces.R
\name{faces}
\alias{faces}
\title{Webmorph Faces}
\usage{
faces(dir = c("composite", "test", "london"), pattern = "\\\\.tem$", ...)
}
\arguments{
\item{dir}{the image directory in extdata to get images from}

\item{pattern}{defaults to all tem files}

\item{...}{Other arguments to pass on to \code{read_tem()}}
}
\value{
webmorph_temlist
}
\description{
A convenience function to get a demo temlist
}
\examples{
faces("test") \%>\% plot()

}
