% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{fileDelete}
\alias{fileDelete}
\title{Delete files on webmorph.org}
\usage{
fileDelete(files)
}
\arguments{
\item{files}{files to delete}
}
\value{
list of deleted files
}
\description{
Delete files on webmorph.org
}
