% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{fileDownload}
\alias{fileDownload}
\title{Download files from webmorph}
\usage{
fileDownload(files, destination = NULL)
}
\arguments{
\item{files}{a list or vector of file names to download, must start with the project number, e.g. "1/averages/f_multi.jpg"}

\item{destination}{A folder to save the files to, defaults to the directory structure of the project}
}
\value{
a list of local paths to the files
}
\description{
Download files from webmorph
}
