% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{fileUpload}
\alias{fileUpload}
\title{Upload files to webmorph.org}
\usage{
fileUpload(files, dir = "/")
}
\arguments{
\item{files}{vector of paths of files to upload}

\item{dir}{directory to upload them to}
}
\value{
list of successfully uploaded files
}
\description{
Upload files to webmorph.org
}
