% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getControlPoints}
\alias{getControlPoints}
\title{Get Control Points for Bezier Curves}
\usage{
getControlPoints(lpath, i = 1)
}
\arguments{
\item{lpath}{line coordinates}

\item{i}{index}
}
\value{
data frame of 4 x and y coordinates
}
\description{
Get Control Points for Bezier Curves
}
\keyword{internal}
