% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_when.R
\name{\%||\%}
\alias{\%||\%}
\title{Conditional Or}
\usage{
a \%||\% b
}
\arguments{
\item{a}{value (if not null)}

\item{b}{value if a is null}
}
\value{
a or (if a is null) b
}
\description{
Conditional Or
}
\examples{
x <- list(b = 2, c = 3)
x$a \%||\% x$b \%||\% x$c \%||\% "default_value"

}
