% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Add a Label to Stimuli}
\usage{
label(
  stimlist,
  text = "LABEL",
  color = "black",
  size = 5,
  alpha = 1,
  angle = 0,
  family = "sans",
  fontface = "plain",
  lineheight = 1.2,
  position = "top center",
  x = NULL,
  y = NULL,
  hjust = NULL,
  vjust = NULL
)
}
\arguments{
\item{stimlist}{list of class webmorph_list}

\item{text}{a vector of the label text(s)}

\item{color}{a vector of the label colour(s)}

\item{size}{a vector of the label size(s)}

\item{alpha}{a vector of the label alpha(s)}

\item{angle}{a vector of the label angle(s) in degrees}

\item{family}{a vector of the label font family(s)}

\item{fontface}{a vector of the label fontface(s)}

\item{lineheight}{a vector of the label lineheight(s)}

\item{position}{vertical (top, middle, bottom) and horizontal (left, center, right) position of the label (e.g., "top left")}

\item{x, y}{a vector of the label x and y position(s), overrides values set from position}

\item{hjust, vjust}{a vector of the label horizontal and vertical justification(s), overrides values set from position}
}
\value{
webmorph_list with labelled images
}
\description{
Add a Label to Stimuli
}
\examples{

faces("test") \%>\%
  label(c("CHINWE", "GEORGE"), color = "red") \%>\%
  plot_fig(labels = "")
}
