% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{login}
\alias{login}
\title{Login to webmorph.org}
\usage{
login(
  email = Sys.getenv("WEBMORPH_EMAIL"),
  password = Sys.getenv("WEBMORPH_PASSWORD")
)
}
\arguments{
\item{email}{The email address associated with the account}

\item{password}{The password for the account}
}
\description{
Login to webmorph.org
}
