% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{makeTrans}
\alias{makeTrans}
\title{Make a Transform}
\usage{
makeTrans(
  trans_img = NULL,
  from_img = NULL,
  to_img = NULL,
  shape = 0,
  color = 0,
  texture = 0,
  outname = tempfile(),
  norm = c("none", "twopoint", "rigid"),
  normpoint = 0:1,
  sample_contours = TRUE,
  warp = c("multiscale", "linear", "multiscalerb"),
  format = c("jpg", "png", "gif")
)
}
\arguments{
\item{trans_img}{image to transform}

\item{from_img}{negative end of the transform dimension}

\item{to_img}{positive end of the transform dimension}

\item{shape}{amount to transform shape}

\item{color}{amount to transform color}

\item{texture}{amount to transform texture}

\item{outname}{local path to save transform to}

\item{norm}{how to normalise}

\item{normpoint}{points for twopoint normalisation}

\item{sample_contours}{whether to sample contours}

\item{warp}{warping algorithm to use}

\item{format}{image format}
}
\value{
webmorph_list
}
\description{
The first 7 arguments are vectorised, so you can put in a vector of image names or shape/color/texture values.
}
