% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.webmorph_list}
\alias{plot.webmorph_list}
\title{Plot webmorph_list}
\usage{
\method{plot}{webmorph_list}(x, y, ...)
}
\arguments{
\item{x}{webmorph_list}

\item{y}{omitted}

\item{...}{Arguments to be passed to \code{\link{plot_fig}}}
}
\value{
plot
}
\description{
Plot webmorph_list
}
\examples{
plot(faces("test"))

}
