% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tem.R
\name{plot.webmorph_temlist}
\alias{plot.webmorph_temlist}
\title{Plot webmorph template list}
\usage{
\method{plot}{webmorph_temlist}(x, y, ...)
}
\arguments{
\item{x}{webmorph_temlist list}

\item{y}{omitted}

\item{...}{Arguments to be passed to ggplot2 (width, height, pt.color, pt.size, pt.shape, bg.color, bg.fill)}
}
\value{
plot
}
\description{
Plot webmorph template list
}
