% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fig}
\alias{plot_fig}
\title{Plot webmorph_list}
\usage{
plot_fig(
  stimlist,
  width = NA,
  height = NA,
  filename = tempfile(fileext = ".png"),
  units = c("in", "cm", "mm"),
  bg.color = "white",
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{stimlist}{webmorph_list}

\item{width}{width of each image (in inches)}

\item{height}{width of each image (in inches)}

\item{filename}{path to save file to}

\item{units}{units for width, height, and margin}

\item{nrow}{number of rows in the figure}

\item{ncol}{number of columns in the figure}

\item{...}{Arguments to be passed to plot_stim}

\item{bc.color}{background colour of plot}
}
\value{
plot
}
\description{
Plot webmorph_list
}
\examples{
faces("test") \%>\% plot(pt.plot = TRUE)

}
