% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stim}
\alias{plot_stim}
\title{Plot webmorph stimulus}
\usage{
plot_stim(
  stim,
  img.plot = TRUE,
  pt.plot = FALSE,
  line.plot = FALSE,
  width = NULL,
  height = NULL,
  label.position = "top center",
  ...
)
}
\arguments{
\item{stim}{webmorph_stim}

\item{img.plot}{Whether to include the image in the plot}

\item{pt.plot}{Whether to include delineation points in the plot}

\item{line.plot}{Whether to include lines in the plot}

\item{width}{plot width}

\item{height}{plot height}

\item{label.position}{vertical (top, middle, bottom) and horizontal (left, center, right) position of the label}

\item{...}{Arguments controlling the plot characteristics}
}
\value{
plot
}
\description{
Arguments for plot characteristics can be:
}
\details{
Points: pt.color, pt.size, pt.shape,
Lines: line.color, line.alpha,
Border: border.color, border.width,
Label: label.text, label.color, label.alpha,
label.size, label.family, label.fontface,
label.x, label.y, hjust, vjust, label.angle
}
\examples{
# default values
plot(faces("test")[[1]],
     width = NULL,  # get from image or tem
     height = NULL, # get from image or tem
     img.plot = TRUE,
     pt.plot = FALSE,
     pt.color = "#00AA00",
     pt.size = 1,
     pt.shape = 3,
     pt.alpha = 1,
     line.plot = FALSE, # T, F, or "bezier"
     line.color = "#FFFF66",
     line.alpha = 0.5
)

}
