% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.webmorph_list}
\alias{print.webmorph_list}
\title{Print webmorph_list}
\usage{
\method{print}{webmorph_list}(x, ...)
}
\arguments{
\item{x}{a list of class webmorph_list}

\item{...}{arguments passed to or from other methods}
}
\value{
prints summary info and returns x
}
\description{
Print webmorph_list
}
