% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.webmorph_stim}
\alias{print.webmorph_stim}
\title{Print webmorph_stim}
\usage{
\method{print}{webmorph_stim}(x, ...)
}
\arguments{
\item{x}{a list of class webmorph_stim}

\item{...}{arguments passed to or from other methods}
}
\value{
prints summary info and returns x
}
\description{
Print webmorph_stim
}
