% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tem.R
\name{print.webmorph_tem}
\alias{print.webmorph_tem}
\title{Print webmorph templates}
\usage{
\method{print}{webmorph_tem}(x, ...)
}
\arguments{
\item{x}{a list of class webmorph_tem}

\item{...}{arguments passed to or from other methods}
}
\value{
prints the template info
}
\description{
Print webmorph templates
}
