% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.webmorph_temlist}
\alias{print.webmorph_temlist}
\title{Print webmorph template list}
\usage{
\method{print}{webmorph_temlist}(x, ...)
}
\arguments{
\item{x}{a list of class webmorph_temlist}

\item{...}{arguments passed to or from other methods}
}
\value{
prints the template info
}
\description{
Print webmorph template list
}
