% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{projListGet}
\alias{projListGet}
\title{Get webmorph.org project list}
\usage{
projListGet(notes = FALSE)
}
\arguments{
\item{notes}{whether to resturn the notes for each project}
}
\value{
data frame of project id, name and notes (optional)
}
\description{
Get webmorph.org project list
}
