% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{projSet}
\alias{projSet}
\title{Set the Working Project}
\usage{
projSet(project)
}
\arguments{
\item{project}{the ID of the project to set}
}
\value{
project id and permissions
}
\description{
Set the Working Project
}
