% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_delete.R
\name{pt_delete}
\alias{pt_delete}
\title{Delete template points}
\usage{
pt_delete(stimlist, ...)
}
\arguments{
\item{stimlist}{list of class webmorph_list}

\item{...}{vectors of points to delete}
}
\value{
webmorph_list with altered templates
}
\description{
Delete template points
}
\examples{
faces("test") \%>\%
  pt_delete(frl_features("mouth")) \%>\%
  plot(pt.plot = TRUE, line.plot = TRUE)

}
