% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stim.R
\name{read_stim}
\alias{read_stim}
\title{Read stimuli}
\usage{
read_stim(path, pattern = NULL, ...)
}
\arguments{
\item{path}{Path to directory containing image and/or template files (or a single file path)}

\item{pattern}{Pattern to use to search for files}

\item{...}{further arguments to pass on to other functions, e.g., breaks argument for unique_names()}
}
\value{
a list of lists with class webmorph_tem
}
\description{
Read stimuli
}
\examples{
path <- system.file("extdata/composite", package = "webmorph")
stimlist <- read_stim(path)
stimlist

}
