% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tem.R
\name{read_tem}
\alias{read_tem}
\title{Read Webmorph template file}
\usage{
read_tem(path, pattern = "\\\\.tem$", images = TRUE, ...)
}
\arguments{
\item{path}{Path to directory containing template files (or a single template file)}

\item{pattern}{Pattern to use to search for files}

\item{images}{Whether to include images}

\item{...}{further arguments to pass on to other functions, e.g., breaks argument for unique_names()}
}
\value{
a list of lists with class webmorph_tem
}
\description{
Read Webmorph template file
}
\examples{
path <- system.file("extdata/composite", package = "webmorph")
temlist <- read_tem(path)

}
