% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename}
\alias{rename}
\title{Rename temlist}
\usage{
rename(
  temlist,
  prefix = "",
  suffix = "",
  new_names = NULL,
  pattern = NULL,
  replacement = NULL,
  ...
)
}
\arguments{
\item{temlist}{A webmorph temlist}

\item{prefix}{String to prefix to each name}

\item{suffix}{String to append to each name}

\item{new_names}{Vector of new names - must be the same length as the temlist}

\item{pattern}{Pattern for gsub}

\item{replacement}{Replacement for gsub}

\item{...}{Additional arguments to pass on to \code{base::gsub()}}
}
\value{
a webmorph temlist
}
\description{
Rename temlist
}
\examples{
faces() \%>\%
  rename(prefix = "new_") \%>\%
  names()
}
