% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep.webmorph_list}
\alias{rep.webmorph_list}
\title{Repeat stim in a list}
\usage{
\method{rep}{webmorph_list}(x, ...)
}
\arguments{
\item{x}{A webmorph_list}

\item{...}{Additional arguments to pass on to \code{base::rep()}}
}
\value{
A webmorph_list
}
\description{
Repeat stim in a list
}
\examples{
faces("test") \%>\%
  rep(3) \%>\%
  rotate(seq(10, 60, 10), fill = rainbow(6)) \%>\%
  plot()
}
