% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep.webmorph_stim}
\alias{rep.webmorph_stim}
\title{Repeat webmorph_stim in a list}
\usage{
\method{rep}{webmorph_stim}(x, ...)
}
\arguments{
\item{x}{A list of class webmorph_stim}

\item{...}{Additional arguments to pass on to \code{base::rep()}}
}
\value{
A webmorph_list
}
\description{
Repeat webmorph_stim in a list
}
\examples{
a <- faces()
rep(a[[1]], 3)
}
