% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep.webmorph_tem}
\alias{rep.webmorph_tem}
\title{Repeat tems in a temlist}
\usage{
\method{rep}{webmorph_tem}(x, ...)
}
\arguments{
\item{x}{A single webmorph template}

\item{...}{Additional arguments to pass on to \code{base::rep()}}
}
\value{
A webmorph temlist
}
\description{
Repeat tems in a temlist
}
\examples{
a <- faces()
rep(a[[1]], 3)
}
