% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep.webmorph_temlist}
\alias{rep.webmorph_temlist}
\title{Repeat items in a temlist}
\usage{
\method{rep}{webmorph_temlist}(x, ...)
}
\arguments{
\item{x}{A webmorph temlist}

\item{...}{Additional arguments to pass on to \code{base::rep()}}
}
\value{
A webmorph temlist
}
\description{
Repeat items in a temlist
}
\examples{
faces("test") \%>\%
  rep(3) \%>\%
  rotate(seq(10, 60, 10), fill = rainbow(6)) \%>\%
  plot()
}
