% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize.R
\name{resize}
\alias{resize}
\title{Title}
\usage{
resize(temlist, width = NULL, height = NULL)
}
\arguments{
\item{temlist}{list of webmorph templates}

\item{width}{new width (in pixels or percent)}

\item{height}{new height (in pixels or percent)}
}
\value{
temlist with resized tems and/or images
}
\description{
Title
}
