% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate templates and images}
\usage{
rotate(temlist, degrees = 0, fill = "none", patch = FALSE)
}
\arguments{
\item{temlist}{list of webmorph templates}

\item{degrees}{degrees to rotate}

\item{fill}{background color}

\item{patch}{whether to use the patch function to set the background color}
}
\value{
temlist with rotated tems and/or images
}
\description{
Rotate templates and images
}
\examples{
faces("test") \%>\%
  rotate(45, fill = "dodgerblue") \%>\%
  plot(pt.plot = TRUE)

faces("test") \%>\%
  rotate(45, patch = TRUE) \%>\%
  plot()

}
