% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.webmorph_list}
\alias{[.webmorph_list}
\title{Extract webmorph_list elements}
\usage{
\method{[}{webmorph_list}(x, i)
}
\arguments{
\item{x}{webmorph_list from which to extract elements}

\item{i}{indices to be selected}
}
\value{
webmorph_list
}
\description{
Extract webmorph_list elements
}
