% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[[<-.webmorph_list}
\alias{[[<-.webmorph_list}
\title{Replace webmorph_list element}
\usage{
\method{[[}{webmorph_list}(x, i) <- value
}
\arguments{
\item{x}{webmorph_list from which to extract elements}

\item{i}{index to be replaced}

\item{value}{webmorph_stim element to replace with}
}
\value{
webmorph_list
}
\description{
Replace webmorph_list element
}
