% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem_def.R
\name{tem_def}
\alias{tem_def}
\title{Get template definition}
\usage{
tem_def(tem_id = 1, path = NULL)
}
\arguments{
\item{tem_id}{}
}
\value{

}
\description{
Get template definition
}
\examples{

frl_face <- tem_def(1)
frl_face$points$name[[19]]
}
