% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tps.R
\name{tems_to_array}
\alias{tems_to_array}
\title{Convert stimlist to array for geomorph}
\usage{
tems_to_array(stimlist)
}
\arguments{
\item{stimlist}{list of class webmorph_list}
}
\value{
3D array
}
\description{
Convert stimlist to array for geomorph
}
\examples{
data <- faces() \%>\% tems_to_array()
dim(data)

}
