% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{transform}
\alias{transform}
\title{Batch Transform  (internal)}
\usage{
transform(
  batch,
  project_id = Sys.getenv("webmorph_project_id"),
  norm = c("none", "twopoint", "rigid"),
  normpoint = 0:1,
  sample_contours = TRUE,
  warp = c("multiscale", "linear", "multiscalerb"),
  format = c("jpg", "png", "gif")
)
}
\arguments{
\item{batch}{checked data frame containing batch info}

\item{project_id}{the project ID}

\item{norm}{how to normalise}

\item{normpoint}{points for twopoint normalisation}

\item{sample_contours}{whether to sample contours}

\item{warp}{warping algorithm to use}

\item{format}{image format}
}
\value{
webmorph_list
}
\description{
Batch Transform  (internal)
}
