% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web}
\alias{web}
\title{Webmorph.org API}
\usage{
web(script = "webmorphR", ..., .error = c("warn", "stop", "none"))
}
\arguments{
\item{script}{the name of the webmorph script}

\item{...}{arguments to pass on to the script in POST}

\item{.error}{whether to warn, stop or do nothing on error}
}
\value{
the response as a list
}
\description{
Webmorph.org API
}
\examples{

web(a = 1, b = FALSE, c = "testing")
}
