% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stim.R
\name{write_stim}
\alias{write_stim}
\title{Write tems and images to files}
\usage{
write_stim(stimlist, dir = ".", prefix = "")
}
\arguments{
\item{stimlist}{list of class webmorph_list}

\item{dir}{Directory to save to}

\item{prefix}{Prefix to add to file names}
}
\value{
list of saved paths
}
\description{
Write tems and images to files
}
