% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tem.R
\name{write_tem}
\alias{write_tem}
\title{Write tems and images to files}
\usage{
write_tem(temlist, dir = ".", prefix = "", suffix = "")
}
\arguments{
\item{temlist}{list of webmorph templates}

\item{dir}{Directory to save to}

\item{prefix}{Prefix to add to file names}

\item{suffix}{Suffix to add to file names}
}
\value{
list of saved paths
}
\description{
Write tems and images to files
}
