% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{average}
\alias{average}
\title{Average templates}
\usage{
average(temlist, name = "average")
}
\arguments{
\item{temlist}{list of webmorph templates}

\item{name}{Name for the average}
}
\value{
webmorph_tem list
}
\description{
Average templates
}
\examples{
faces("london") \%>\%
  average() \%>\%
  plot(img.plot = FALSE, pt.plot = TRUE, line.plot = TRUE)

}
