% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{Crop images and templates}
\usage{
crop(
  temlist,
  width = NULL,
  height = NULL,
  x_off = 0,
  y_off = 0,
  fill = "white"
)
}
\arguments{
\item{temlist}{list of webmorph templates}

\item{width}{width of cropped image}

\item{height}{height of cropped image}

\item{x_off}{x-offset}

\item{y_off}{y-offset}

\item{fill}{background colour if cropping goes outside the original image}
}
\value{
temlist with cropped tems and/or images
}
\description{
Arguments width, height, x_off and y_off can bet set in pixels or proportions (between 0.0 and 2.0)
}
\examples{
path <- system.file("extdata/composite/f_multi.tem", package = "webmorph")
read_tem(path) \%>\%
  crop(.5, .5, .25, .25) \%>\%
  plot(image = TRUE)

}
