% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces.R
\name{faces}
\alias{faces}
\title{Webmorph Faces}
\usage{
faces(dir = c("composite", "test", "london", "lisa"), pattern = NULL, ...)
}
\arguments{
\item{dir}{the directory in extdata to get files from}

\item{pattern}{defaults to all files}

\item{...}{Other arguments to pass on to \code{read_tem()}}
}
\value{
webmorph_list
}
\description{
A convenience function to get a demo stimlist
}
\examples{
faces("test") \%>\% plot()

}
