% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{getProjectID}
\alias{getProjectID}
\title{Get Project ID from a list of filenames}
\usage{
getProjectID(files)
}
\arguments{
\item{files}{}
}
\value{
project ID
}
\description{
Get Project ID from a list of filenames
}
