% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{makeAvg}
\alias{makeAvg}
\title{Make an Average Face}
\usage{
makeAvg(
  files,
  outname = tempfile(),
  texture = TRUE,
  norm = c("none", "twopoint", "rigid"),
  normpoint = 0:1,
  format = c("jpg", "png", "gif")
)
}
\arguments{
\item{files}{the image files to average}

\item{outname}{local path to save average to}

\item{texture}{logical, textured average}

\item{norm}{how to normalise}

\item{normpoint}{points for twopoint normalisation}

\item{format}{image format}
}
\value{
webmorph_list
}
\description{
Make an Average Face
}
