% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.webmorph_stim}
\alias{plot.webmorph_stim}
\title{Plot webmorph stimulus}
\usage{
\method{plot}{webmorph_stim}(x, y, ...)
}
\arguments{
\item{x}{webmorph_stim list}

\item{y}{omitted}

\item{...}{Arguments controlling the plot characteristics (img.plot, pt.plot, line.plot, width, height, pt.color, pt.size, pt.shape, line.color, line.alpha, border.color, border.width)}
}
\value{
plot
}
\description{
Plot webmorph stimulus
}
\examples{
# default values
plot(faces("test")[[1]],
     width = NULL,  # get from image or tem
     height = NULL, # get from image or tem
     img.plot = TRUE,
     pt.plot = FALSE,
     pt.color = "#00AA00",
     pt.size = 1,
     pt.shape = 3,
     pt.alpha = 1,
     line.plot = FALSE, # T, F, or "bezier"
     line.color = "#FFFF66",
     line.alpha = 0.5
)

#custom settings
plot(faces("test")[[1]],
     img.plot = FALSE,
     pt.plot = TRUE,
     pt.color = "dodgerblue",
     pt.shape = 1,
     pt.alpha = 0.5,
     line.plot = TRUE,
     border.color = "black",
     border.width = 10 # in pixels
)

}
