% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.webmorph_stim}
\alias{plot.webmorph_stim}
\title{Plot webmorph_stim}
\usage{
\method{plot}{webmorph_stim}(x, y, ...)
}
\arguments{
\item{x}{webmorph_stim}

\item{y}{omitted}

\item{...}{Arguments to be passed to \code{\link{plot_stim}}}
}
\value{
ggplot
}
\description{
Plot webmorph_stim
}
\examples{
plot(faces("test")[[1]])
}
