% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_delete.R
\name{pt_delete}
\alias{pt_delete}
\title{Delete template points}
\usage{
pt_delete(temlist, ...)
}
\arguments{
\item{temlist}{list of webmorph templates}

\item{...}{vectors of points to delete}
}
\value{
temlist with altered tems
}
\description{
Delete template points
}
