% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize.R
\name{resize}
\alias{resize}
\title{Resize templates and images}
\usage{
resize(stimlist, width = NULL, height = NULL)
}
\arguments{
\item{stimlist}{list of class webmorph_list}

\item{width}{new width (in pixels or percent)}

\item{height}{new height (in pixels or percent)}
}
\value{
webmorph_list with resized tems and/or images
}
\description{
Resize templates and images
}
\examples{
faces("test") \%>\%
  resize(.5, .75) \%>\%
  plot(pt.plot = TRUE,
       pt.color = "black",
       pt.shape = 1)

}
