% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate templates and images}
\usage{
rotate(temlist, degrees = 0)
}
\arguments{
\item{temlist}{list of webmorph templates}

\item{degrees}{degrees to rotate}
}
\value{
temlist with rotated tems and/or images
}
\description{
Rotate templates and images
}
\examples{
path <- system.file("extdata/composite/f_multi.tem", package = "webmorph")
read_tem(path) \%>\%
  rotate(45) \%>\%
  plot(image = TRUE)

}
