% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tps.R
\name{tems_to_array}
\alias{tems_to_array}
\title{Convert temlist to array for geomorph}
\usage{
tems_to_array(temlist)
}
\arguments{
\item{temlist}{list of webmorph templates}
}
\value{
3D array
}
\description{
Convert temlist to array for geomorph
}
\examples{
data <- faces() \%>\% tems_to_array()
dim(data)

}
