% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tps.R
\name{write_tps}
\alias{write_tps}
\title{Create a TPS file from a webmorph_list}
\usage{
write_tps(stimlist, path_to_tps = NULL)
}
\arguments{
\item{stimlist}{list of class webmorph_list}

\item{path_to_tps}{optional filename to save TPS file}
}
\value{
text of tps file
}
\description{
Create a TPS file from a webmorph_list
}
\examples{
# set path_to_tps to save to a file
faces("test") \%>\%
  write_tps() \%>\%
  cat()

}
