% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tps.R
\name{write_tps}
\alias{write_tps}
\title{Create a TPS file from a list of webmorph templates}
\usage{
write_tps(temlist, path_to_tps = NULL)
}
\arguments{
\item{temlist}{list of webmorph templates}
}
\value{
text of tps file
}
\description{
Create a TPS file from a list of webmorph templates
}
\examples{
# set path_to_tps to save to a file
faces("test") \%>\%
  write_tps() \%>\%
  cat()

}
